Fix parsing when pdf gets messed up a bit
Sometimes it happens that week section is in the middle of the menus section. The new day identification will find the id of the current day based on the date, it does not rely on the menu position in text. Otherwise, if all else fails, the pdf links are returned.